+++ ZenUnitUtilsAndAssertionTests/Components/Time/MetalMock/StopwatchMock.h
@@ -4,6 +4,6 @@ class StopwatchMock : public Metal::Mock<ZenUnit::Stopwatch>
 {
 public:
    METALMOCK_VOID0(Start)
-   METALMOCK_NONVOID0(long long, GetElapsedMicrosecondsThenResetStopwatch)
+   METALMOCK_NONVOID0(unsigned, GetElapsedMicrosecondsThenResetStopwatch)
    METALMOCK_NONVOID0(string, StopAndGetElapsedSeconds)
 };
